% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_tools.R
\name{add2json}
\alias{add2json}
\title{Add Data to JSON file}
\usage{
add2json(list, json_filename, jsonpath = NULL)
}
\arguments{
\item{list}{A list with extra information}

\item{json_filename}{The json file to be supplemented. Preferably
without file extensions. Can contain the full path.}

\item{jsonpath}{The path for the json file. Only required if it
has not been provided in the json_filename.}
}
\value{
Outputs the supplemented json file.
}
\description{
Add Data to JSON file
}
\examples{
add2json(list1, 'pet/sub-01_pet')

}
