% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_outputs.R
\name{ancmat_getData}
\alias{ancmat_getData}
\title{Creates a BIDS list from a pipeline anc.mat file}
\usage{
ancmat_getData(matfile)
}
\arguments{
\item{matfile}{The anc.mat file generated by the pipeline.}
}
\value{
A list in BIDS structure
}
\description{
This function creates a BIDS list from a pipeline anc file.
}
\examples{
\dontrun{
ancmat_getData('ancfile.mat')
}
}
