% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_checks.R
\name{check_dcm2niix}
\alias{check_dcm2niix}
\title{Check if dcm2niix is installed and in the path}
\usage{
check_dcm2niix()
}
\value{
Logical of whether dcm2niix was successfully called from the path.
}
\description{
Check if dcm2niix is installed and in the path
}
\examples{
check_dcm2niix()

}
