% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_checks.R
\name{check_tpcclib}
\alias{check_tpcclib}
\title{Check if tpcclib is installed and in the path}
\usage{
check_tpcclib()
}
\value{
Logical of whether tpcclib was successfully called from the path.
}
\description{
Check if tpcclib is installed and in the path
}
\examples{
check_tpcclib()

}
