% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_checks.R
\name{check_winlinux}
\alias{check_winlinux}
\title{Check if the Windows Ubuntu extension is installed and available}
\usage{
check_winlinux()
}
\value{
Logical of whether Ubuntu can be called from Windows
}
\description{
Check if the Windows Ubuntu extension is installed and available
}
\examples{
check_winlinux()

}
