% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_tools.R
\name{combine_jsons}
\alias{combine_jsons}
\title{Combine JSON files}
\usage{
combine_jsons(json_filenames, out_filename, outpath = getwd())
}
\arguments{
\item{json_filenames}{Vector of the json files to be combined. Preferably
without file extensions. Preferably containing the full path.}

\item{out_filename}{Filename of the output json file.  Preferably without
file extensions.}

\item{outpath}{Path to the output files. Defaults to the working directory.}

\item{inpath}{Path to the input files. Defaults to the working directory.}
}
\value{
Creates a combined .json file.
}
\description{
Combine JSON files
}
\examples{
k <- list(abc = 1, def = 2)
j <- list(abcd = 3, defg = 4)
jsonlite::write_json(k, 'json1.json')
jsonlite::write_json(j, 'json2.json')

combine_jsons(c('json1', 'json2'), 'combined')
}
