% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_convert.R
\name{dicom_attributes}
\alias{dicom_attributes}
\title{Get Attributes of DICOM files}
\usage{
dicom_attributes(inpath = getwd())
}
\arguments{
\item{inpath}{The path where the DICOM files are located.  A file search will
take place in all nested directories. This allows for convenience in
directory specification, but be careful about having lots of files in
nested places within the specified folder.}
}
\value{
A tibble of the details and attributes of the found dcm files.
}
\description{
This function allows users to select which DICOM folder to convert for
whichever purpose.  Heuristic strategies can be used to select from this list
}
\examples{
dicom_attributes()
}
