% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_convert.R
\name{ecat2nii}
\alias{ecat2nii}
\title{Ecat7 to NIfTI Converter}
\usage{
ecat2nii(v_filename, inpath = getwd(), out_filename = NULL,
  outpath = getwd(), checkLines = F, compressFile = T)
}
\arguments{
\item{v_filename}{Filename of the input .v file. Preferably without file
extension. Should only be the filename. Path goes into inpath.}

\item{inpath}{Path to the input file. Defaults to working directory.}

\item{out_filename}{Filename of the output file. Preferably without file
extension. Defaults to same as the input filename. Should only be the
filename. Path goes into outpath.}

\item{outpath}{Path where the output file should be placed. Defaults to the
working directory.}

\item{checkLines}{Should the system commands be checked (and not run)?
Default FALSE.}

\item{compressFile}{Should the output be compressed as a .nii.gz file?
Defaulse TRUE.}
}
\value{
If checkLines is TRUE, the commands will be returned.  If checkLines
is FALSE, the binary success of each command will be returned.
}
\description{
Wrapper for the Turku PET Centre ecat2nii function
}
\examples{
ecat2nii('r_abcd_1', out_filename = 'abcd_1', checkLines = F, compressFile = T)

}
