% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_convert.R
\name{ecat_info}
\alias{ecat_info}
\title{Extract information from Ecat7 files}
\usage{
ecat_info(v_filename, inpath = getwd(), outpath = getwd(), checkLines = F)
}
\arguments{
\item{v_filename}{Filename of the input .v file. Preferably without file
extension (except if other dots in the filename). Should only be the
filename. Path goes into inpath.}

\item{inpath}{Path to the input file. Defaults to the working directory.}

\item{outpath}{Path to which data can be written. Just required for temporary
files. Defaults to the working directory.}

\item{checkLines}{Should the system commands be checked (and not run)?
Default FALSE.}
}
\value{
If checkLines is TRUE, the commands will be returned.  If checkLines
is FALSE, the relevant data will be returned in a list.
}
\description{
Extracts header, decay, counts and timing information from ecat7 .v files.
}
\examples{
ecat_info('r_abcd_1', checkLines = F)

}
