% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_tools.R
\name{ecat_info2bids_json}
\alias{ecat_info2bids_json}
\title{Extract Information from ecat7 Header into BIDS Sidecar}
\usage{
ecat_info2bids_json(v_filename, out_filename, inpath = getwd(),
  outpath = getwd())
}
\arguments{
\item{v_filename}{Filename of the input .v file. Preferably without file
extension. Should only be the filename. Path goes into inpath.}

\item{out_filename}{Filename of the output file. Preferably without file
extension. Should only be the filename. Path goes into outpath.}

\item{inpath}{Path to the input file. Defaults to the working directory.}

\item{outpath}{Path to the output file. Defaults to the working directory.}
}
\value{
Produces the PET BIDS sidecar with all relevant information from the
.v file
}
\description{
Extract Information from ecat7 Header into BIDS Sidecar
}
\examples{
ecat_info2bids_json('abcd_1', 'sub-01_ses-2015_task-rest_acq-T807AV1451_pet')

}
