% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_checks.R
\name{fix_command_linux}
\alias{fix_command_linux}
\title{Fix a command such that it can run on Windows Ubuntu}
\usage{
fix_command_linux(command_text)
}
\arguments{
\item{command_text}{One or more commands as characters which will be appropriately prefixed.}
}
\value{
The texts with the correct prefixes.
}
\description{
Fix a command such that it can run on Windows Ubuntu
}
\examples{
fix_command_linux('uname -a')
}
