% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_checks.R
\name{get_os}
\alias{get_os}
\title{Get the OS}
\usage{
get_os()
}
\value{
The OS: win/mac/unix
}
\description{
Note: This function is copied directly from https://github.com/r-lib/rappdirs/blob/master/R/utils.r#L1
}
\examples{
get_os()

}
\references{
https://github.com/r-lib/rappdirs/blob/master/R/utils.r#L1
}
