% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_convert.R
\name{get_studydb_data}
\alias{get_studydb_data}
\title{Get studydb data for a single subject's folder}
\usage{
get_studydb_data(subjFolder, path = getwd())
}
\arguments{
\item{subjFolder}{The name of the subject folder.}

\item{path}{The file path of the study folder.}
}
\value{
A data.frame with relevant information from the studyDB file.
}
\description{
For doing this for a whole study folder, use the \emph{get_studydb_data_folder} command
}
\examples{
abcd_dat <- get_studydb_data('abcd')

}
