% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelinederivative_outputs.R
\name{getsrtmresults}
\alias{getsrtmresults}
\title{Extract outcome parameters from an srtmresults text file}
\usage{
getsrtmresults(filename, skipno)
}
\arguments{
\item{filename}{Filename of the txt file}

\item{skipno}{The number of rows to skip (from getsrtmresults_skiprows)}
}
\value{
The outcome parameters
}
\description{
Convenience function for srtm2xlsx.  Extract the outcome parameters from a
srtmresults file.
}
\examples{
getsrtmresults('abcd_1_kinfitresults.txt', 15)

}
