% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_outputs.R
\name{kfresults_getData}
\alias{kfresults_getData}
\title{Extract TAC data from kinfitresults}
\usage{
kfresults_getData(matfile)
}
\arguments{
\item{matfile}{Filename of the kinfitresults file}
}
\value{
List with the sizes of all rois, the subject name, PET number, and
TAC data
}
\description{
Extract TAC data from a kinfitresults.mat file
}
\examples{
abcd_out <- kfresults_getData('abcd_1_kinfitresults.mat')

}
