% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kipettools-package.R
\docType{package}
\name{kipettools-package}
\alias{kipettools-package}
\alias{_PACKAGE}
\title{kipettools package}
\description{
KI PET Group Tools for R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mathesong/kipettools}
  \item Report bugs at \url{https://github.com/mathesong/kipettools/issues}
}

}
\author{
Granville Matheson \href{mailto:mathesong@gmail.com}{mathesong@gmail.com}
}
\keyword{internal}
