% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_convert.R
\name{nicebracketer}
\alias{nicebracketer}
\title{Nice Bracketer}
\usage{
nicebracketer(filelocation)
}
\arguments{
\item{filelocation}{A path name}
}
\value{
The same path name, but with forward slashes
}
\description{
Because Windows brackets look ugly for folder locations
}
\examples{
nicebracketer('C:\\\\Users\\\\User\\\\Documents\\\\SpecialFile.docx')
nicebracketer('C:\\\\Users\\\\User\\\\Documents')


}
