% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_convert.R
\name{read_kgblood}
\alias{read_kgblood}
\title{Read blood data from Kaleidagraph output}
\usage{
read_kgblood(filename, inpath = getwd())
}
\arguments{
\item{filename}{Name of the file including extension, but excluding path. This function only works for .txt files, and not for Kaleidagraph's qda format. Files must be saved as tab-separated txt files in Kaleidagraph before they can be used with this function.}

\item{inpath}{Path to the input file. Defaults to the working directory.}
}
\value{
List containing the automatic and manual blood data
}
\description{
Read blood data from Kaleidagraph output
}
\examples{
read_kgblood('blood_file.txt')
}
