% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_outputs.R
\name{roistats_getData}
\alias{roistats_getData}
\title{Extract TAC data from roistats}
\usage{
roistats_getData(matfile)
}
\arguments{
\item{matfile}{Filename of the roistats file}
}
\value{
List with the sizes of all rois, the subject name, PET number, and
TAC data
}
\description{
Extract TAC data from a kinfitresults.mat file
}
\examples{
abcd_out <- roistats_getData('abcd_dynpet1_roistats.mat')

}
