% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_convert.R
\name{terminal_munge}
\alias{terminal_munge}
\title{Munge terminal output}
\usage{
terminal_munge(terminal_output, matchpattern = NULL, splitpattern = NULL,
  colnames = NULL, has_header = T)
}
\arguments{
\item{terminal_output}{Raw terminal output. Make sure that it consists of one
data rectangle.}

\item{matchpattern}{Regex pattern for dividing the data set. Uses
stringr::str_match. Specify either matchpattern or splitpattern.}

\item{splitpattern}{Regex pattern for splitting the data set. Uses
stringr::str_split. Specify either splitpattern or matchpattern.}

\item{colnames}{What should the column names of the tibble? If the terminal
output is a list of attributes and their values, then it will be handled
automatically.}

\item{has_header}{Does the table have a header inside it. Default is TRUE.}
}
\value{
A list including a tibble of the output, and, if the output consists
of attributes and values, then it will also include a list.
}
\description{
Split and produce data from terminal output
}
\examples{
hdr_dat <- terminal_munge(outcome_hdr, matchpattern='(.*)\\\\:\\\\= (.*)')

}
