% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_convert.R
\name{get_studydb_data_folder}
\alias{get_studydb_data_folder}
\title{Get studydb data for a study folder}
\usage{
get_studydb_data_folder(studyFolder = getwd(), savecsv = F)
}
\arguments{
\item{studyFolder}{The folder location of the study folder. Default is
current directory.}

\item{savecsv}{Should a csv file be saved? Default is FALSE. If you wish to
save a csv file, replace this argument with the desired filename.}
}
\value{
A data.frame with the data from the studydb files.
}
\description{
Extract studydb data for a whole folder of subjects
}
\examples{
get_studydb_data_folder(savecsv = 'studydata.csv')

}
