% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kipettools-package.R
\docType{package}
\name{kipettools-package}
\alias{kipettools}
\alias{kipettools-package}
\title{kipettools: R Tools For KI PET Group Files}
\description{
This package contains tools for interacting with KI PET group pipeline files with R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mathesong/kipettools}
  \item Report bugs at \url{https://github.com/mathesong/kipettools/issues}
}

}
\author{
\strong{Maintainer}: Granville Matheson \email{mathesong@gmail.com}

}
\keyword{internal}
